ci(xtask): split out github helper functions
